Restructure aiida-plugin into config and main, and separate the functionality out
